+++ tests/Dummies/ExampleModuleServiceProvider.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the ExampleModuleServiceProvider class.
+ *
+ * @copyright   Copyright (c) 2021 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2021-11-13
+ *
+ */
+
+namespace Konekt\Concord\Tests\Dummies;
+
+use Konekt\Concord\BaseModuleServiceProvider;
+
+class ExampleModuleServiceProvider extends BaseModuleServiceProvider
+{
+}