+++ tests/Dummies/manifest_with_name_and_version.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'name' => 'Elmo Song Module',
+    'version' => '7.9.0',
+];