+++ tests/Feature/Module/VersionTest.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the VersionTest class.
+ *
+ * @copyright   Copyright (c) 2021 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2021-11-13
+ *
+ */
+
+namespace Konekt\Concord\Tests\Feature\Module;
+
+use Konekt\Concord\Tests\TestCase;
+
+class VersionTest extends TestCase
+{
+    /** @test */
+    public function the_version_is_not_mandatory()
+    {
+
+    }
+}